Add files via upload
This data looks at the Streaming History of a Spotify user. This data gives information that Spotify can use to target their audience to use the app more often, nudge the consumer to use the app differently, encourage exploration of different artist etc. This workbook includes visualizations and general information that can help the business understand this user on a deeper level.